ACM Project Portion
I added the 3 paragraphs and found a large literal placeholder image to
add underneath. However, I'm not sure how the whole background slides
mechanic works. Only a bit of the image can be seen before the slide
ends and the rest of the site continues. I'll go ahead and send how it
is now, but I'll update again if I figure out the image part. Also added
some CSS at the very bottom for my section. Labeled in the comments